⌨️  type refactor 1
